Many bugs were fixed.
We did the show tonight. There were a few bugs. These are now fixed!:
* Sometimes players can buzz in after the answer has been revealed (will
need further testing on round 4 to confirm)
* Round 3 adds currently selected to the first resolved group even if
it's not supposed to be there when the time runs out.
* Debug goes with default team name instead of actual one first time.
* The round 4 timer should be paused if a team is answering.
* There's no "out of time" sound effect.
* The bounding boxes for the heiroglyphs in rounds 1 and 2 are a bit
iffy.
* The Konami Code prints "KONAMICODE" as you type it and clears if you
get it wrong so you know if it's going to happen.
* The game does weird things the second time it plays through without
quitting (so I've forced it to quit muhuhaha)
* There's no way to manually change score (NOW: q for +1 to first team,
w for -1, o for +1 to first, p for -1).

Also there are some exciting new things to work on:
* WORD WRAP
* Fancy fadey colours in round 3.
* (irrelevant to the code) HARDWARE