A few useful things!
Updated the README!
Added a calibration function at the start of the game which allows you
to set what keypress is the buzzer for each team!
Added some validation for filenames so it doesn't crash when given a bad
one!
Yay me!
(also: I made exciting new hardware, which is cool)